- refactor: duplicated MainActivity code into shared android apps code
